Refactor Legend to GuideRenderer subclass in python
