+++ bokeh/models/__init__.py
@@ -9,6 +9,7 @@ from .callbacks import *
 from .formatters import *
 from .glyphs import *
 from .grids import *
+from .guides import *
 from .layouts import *
 from .images import *
 from .map_plots import *
