+++ bokeh/models/plots.py
@@ -19,10 +19,11 @@ from ..core.properties import (
     List, Dict, Include, Override, TitleProp)
 from ..util.string import nice_join
 
-from .annotations import Legend, Title
+from .annotations import Title
 from .axes import Axis
 from .glyphs import Glyph
 from .grids import Grid
+from .guides import Legend
 from .ranges import Range, FactorRange
 from .renderers import Renderer, GlyphRenderer, DataRenderer, TileRenderer, DynamicImageRenderer
 from .sources import DataSource, ColumnDataSource
