+++ bokeh/models/tests/test_glyphs.py
@@ -1,5 +1,11 @@
 from __future__ import absolute_import
 
+from .utils.property_utils import (
+    FILL, LINE, TEXT, GLYPH, MARKER,
+    check_properties_existence, check_fill_properties,
+    check_line_properties, check_text_properties, check_marker_properties
+)
+
 from bokeh.models.glyphs import (
     AnnularWedge, Annulus, Arc,
     Bezier,
@@ -49,49 +55,6 @@ from bokeh.core.enums import (
     ButtonType, MapType,
     Color, NamedIcon)
 
-FILL = ["fill_color", "fill_alpha"]
-LINE = ["line_color", "line_width", "line_alpha", "line_join", "line_cap", "line_dash", "line_dash_offset"]
-TEXT = ["text_font", "text_font_size", "text_font_style", "text_color", "text_alpha", "text_align", "text_baseline"]
-
-PROPS = ["name", "tags"]
-GLYPH = ["visible"]
-MARKER = ["x", "y", "size", "angle", "angle_units"]
-
-def check_props(glyph, *props):
-    expected = set(sum((PROPS, GLYPH) + props, []))
-    found = set(glyph.properties())
-    missing = expected.difference(found)
-    extra = found.difference(expected)
-    assert len(missing) == 0, "Properties missing: {0}".format(", ".join(sorted(missing)))
-    assert len(extra) == 0, "Extra properties: {0}".format(", ".join(sorted(extra)))
-
-def check_fill(glyph):
-    assert glyph.fill_color == Color.gray
-    assert glyph.fill_alpha == 1.0
-
-def check_line(glyph):
-    assert glyph.line_color == Color.black
-    assert glyph.line_width == 1
-    assert glyph.line_alpha == 1.0
-    assert glyph.line_join == LineJoin.miter
-    assert glyph.line_cap == LineCap.butt
-    assert glyph.line_dash == []
-    assert glyph.line_dash_offset == 0
-
-def check_text(glyph):
-    assert glyph.text_font == "helvetica"
-    assert glyph.text_font_size == {"value": "12pt"}
-    assert glyph.text_font_style == FontStyle.normal
-    assert glyph.text_color == "#444444"
-    assert glyph.text_alpha == 1.0
-    assert glyph.text_align == TextAlign.left
-    assert glyph.text_baseline == TextBaseline.bottom
-
-def check_marker(marker):
-    assert marker.x is None
-    assert marker.y is None
-    assert marker.size == 4
-
 def test_AnnularWedge():
     glyph = AnnularWedge()
     assert glyph.x is None
@@ -101,9 +64,9 @@ def test_AnnularWedge():
     assert glyph.start_angle is None
     assert glyph.end_angle is None
     assert glyph.direction == "anticlock"
-    yield check_fill, glyph
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_fill_properties, glyph
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "x",
         "y",
         "inner_radius",
@@ -115,7 +78,7 @@ def test_AnnularWedge():
         "end_angle",
         "end_angle_units",
         "direction",
-    ], FILL, LINE)
+    ], FILL, LINE, GLYPH)
 
 def test_Annulus():
     glyph = Annulus()
@@ -123,16 +86,16 @@ def test_Annulus():
     assert glyph.y is None
     assert glyph.inner_radius is None
     assert glyph.outer_radius is None
-    yield check_fill, glyph
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_fill_properties, glyph
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "x",
         "y",
         "inner_radius",
         "inner_radius_units",
         "outer_radius",
         "outer_radius_units",
-    ], FILL, LINE)
+    ], FILL, LINE, GLYPH)
 
 def test_Arc():
     glyph = Arc()
@@ -142,8 +105,8 @@ def test_Arc():
     assert glyph.start_angle is None
     assert glyph.end_angle is None
     assert glyph.direction == "anticlock"
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "x",
         "y",
         "radius",
@@ -153,7 +116,7 @@ def test_Arc():
         "end_angle",
         "end_angle_units",
         "direction",
-    ], LINE)
+    ], LINE, GLYPH)
 
 def test_Bezier():
     glyph = Bezier()
@@ -165,8 +128,8 @@ def test_Bezier():
     assert glyph.cy0 is None
     assert glyph.cx1 is None
     assert glyph.cy1 is None
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "x0",
         "y0",
         "x1",
@@ -175,7 +138,7 @@ def test_Bezier():
         "cy0",
         "cx1",
         "cy1",
-    ], LINE)
+    ], LINE, GLYPH)
 
 def test_Gear():
     glyph = Gear()
@@ -187,9 +150,9 @@ def test_Gear():
     assert glyph.pressure_angle == 20
     assert glyph.shaft_size == 0.3
     assert glyph.internal == False
-    yield check_fill, glyph
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_fill_properties, glyph
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "x",
         "y",
         "angle",
@@ -199,7 +162,7 @@ def test_Gear():
         "pressure_angle",
         "shaft_size",
         "internal",
-    ], FILL, LINE)
+    ], FILL, LINE, GLYPH)
 
 def test_Image():
     glyph = Image()
@@ -209,7 +172,7 @@ def test_Image():
     assert glyph.dw is None
     assert glyph.dh is None
     assert glyph.dilate == False
-    yield (check_props, glyph, [
+    yield (check_properties_existence, glyph, [
         "image",
         "x",
         "y",
@@ -219,7 +182,7 @@ def test_Image():
         "dh_units",
         "dilate",
         "color_mapper",
-    ])
+    ], GLYPH)
 
 def test_ImageRGBA():
     glyph = ImageRGBA()
@@ -231,7 +194,7 @@ def test_ImageRGBA():
     assert glyph.rows == None
     assert glyph.cols == None
     assert glyph.dilate == False
-    yield (check_props, glyph, [
+    yield (check_properties_existence, glyph, [
         "image",
         "x",
         "y",
@@ -242,7 +205,7 @@ def test_ImageRGBA():
         "rows",
         "cols",
         "dilate",
-    ])
+    ], GLYPH)
 
 def test_ImageURL():
     glyph = ImageURL()
@@ -257,7 +220,7 @@ def test_ImageURL():
     assert glyph.retry_attempts == 0
     assert glyph.retry_timeout == 0
     assert glyph.global_alpha == 1.0
-    yield (check_props, glyph, [
+    yield (check_properties_existence, glyph, [
         "url",
         "x",
         "y",
@@ -272,27 +235,27 @@ def test_ImageURL():
         "retry_attempts",
         "retry_timeout",
         "global_alpha",
-    ])
+    ], GLYPH)
 
 def test_Line():
     glyph = Line()
     assert glyph.x is None
     assert glyph.y is None
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "x",
         "y",
-    ], LINE)
+    ], LINE, GLYPH)
 
 def test_MultiLine():
     glyph = MultiLine()
     assert glyph.xs is None
     assert glyph.ys is None
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "xs",
         "ys",
-    ], LINE)
+    ], LINE, GLYPH)
 
 def test_Oval():
     glyph = Oval()
@@ -301,9 +264,9 @@ def test_Oval():
     assert glyph.width is None
     assert glyph.height is None
     assert glyph.angle == 0
-    yield check_fill, glyph
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_fill_properties, glyph
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "x",
         "y",
         "width",
@@ -312,29 +275,29 @@ def test_Oval():
         "height_units",
         "angle",
         "angle_units",
-    ], FILL, LINE)
+    ], FILL, LINE, GLYPH)
 
 def test_Patch():
     glyph = Patch()
     assert glyph.x is None
     assert glyph.y is None
-    yield check_fill, glyph
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_fill_properties, glyph
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "x",
         "y",
-    ], FILL, LINE)
+    ], FILL, LINE, GLYPH)
 
 def test_Patches():
     glyph = Patches()
     assert glyph.xs is None
     assert glyph.ys is None
-    yield check_fill, glyph
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_fill_properties, glyph
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "xs",
         "ys",
-    ], FILL, LINE)
+    ], FILL, LINE, GLYPH)
 
 def test_Quad():
     glyph = Quad()
@@ -342,14 +305,14 @@ def test_Quad():
     assert glyph.right is None
     assert glyph.bottom is None
     assert glyph.top is None
-    yield check_fill, glyph
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_fill_properties, glyph
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "left",
         "right",
         "bottom",
         "top",
-    ], FILL, LINE)
+    ], FILL, LINE, GLYPH)
 
 def test_Quadratic():
     glyph = Quadratic()
@@ -359,15 +322,15 @@ def test_Quadratic():
     assert glyph.y1 is None
     assert glyph.cx is None
     assert glyph.cy is None
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "x0",
         "y0",
         "x1",
         "y1",
         "cx",
         "cy",
-    ], LINE)
+    ], LINE, GLYPH)
 
 def test_Ray():
     glyph = Ray()
@@ -375,15 +338,15 @@ def test_Ray():
     assert glyph.y is None
     assert glyph.angle is None
     assert glyph.length is None
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "x",
         "y",
         "angle",
         "angle_units",
         "length",
         "length_units",
-    ], LINE)
+    ], LINE, GLYPH)
 
 def test_Rect():
     glyph = Rect()
@@ -393,9 +356,9 @@ def test_Rect():
     assert glyph.height is None
     assert glyph.angle == 0
     assert glyph.dilate == False
-    yield check_fill, glyph
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_fill_properties, glyph
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "x",
         "y",
         "width",
@@ -405,7 +368,7 @@ def test_Rect():
         "angle",
         "angle_units",
         "dilate",
-    ], FILL, LINE)
+    ], FILL, LINE, GLYPH)
 
 def test_Segment():
     glyph = Segment()
@@ -413,13 +376,13 @@ def test_Segment():
     assert glyph.y0 is None
     assert glyph.x1 is None
     assert glyph.y1 is None
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "x0",
         "y0",
         "x1",
         "y1"
-    ], LINE)
+    ], LINE, GLYPH)
 
 def test_Text():
     glyph = Text()
@@ -427,8 +390,8 @@ def test_Text():
     assert glyph.y is None
     assert glyph.text == "text"
     assert glyph.angle == 0
-    yield check_text, glyph
-    yield (check_props, glyph, [
+    yield check_text_properties, glyph
+    yield (check_properties_existence, glyph, [
         "x",
         "y",
         "text",
@@ -436,7 +399,7 @@ def test_Text():
         "angle_units",
         "x_offset",
         "y_offset"
-    ], TEXT)
+    ], TEXT, GLYPH)
 
 def test_Wedge():
     glyph = Wedge()
@@ -446,9 +409,9 @@ def test_Wedge():
     assert glyph.start_angle is None
     assert glyph.end_angle is None
     assert glyph.direction == "anticlock"
-    yield check_fill, glyph
-    yield check_line, glyph
-    yield (check_props, glyph, [
+    yield check_fill_properties, glyph
+    yield check_line_properties, glyph
+    yield (check_properties_existence, glyph, [
         "x",
         "y",
         "radius",
@@ -458,100 +421,100 @@ def test_Wedge():
         "end_angle",
         "end_angle_units",
         "direction",
-    ], FILL, LINE)
+    ], FILL, LINE, GLYPH)
 
 def test_Asterisk():
     marker = Asterisk()
-    yield check_marker, marker
-    yield check_fill, marker
-    yield check_line, marker
-    yield check_props, marker, MARKER, FILL, LINE
+    yield check_marker_properties, marker
+    yield check_fill_properties, marker
+    yield check_line_properties, marker
+    yield check_properties_existence, marker, MARKER, FILL, LINE, GLYPH
 
 def test_Circle():
     marker = Circle()
-    yield check_marker, marker
+    yield check_marker_properties, marker
     assert marker.radius == None
-    yield check_fill, marker
-    yield check_line, marker
-    yield (check_props, marker, [
+    yield check_fill_properties, marker
+    yield check_line_properties, marker
+    yield (check_properties_existence, marker, [
         "radius",
         "radius_units",
         "radius_dimension",
-    ], MARKER, FILL, LINE)
+    ], MARKER, FILL, LINE, GLYPH)
 
 def test_CircleCross():
     marker = CircleCross()
-    yield check_marker, marker
-    yield check_fill, marker
-    yield check_line, marker
-    yield check_props, marker, MARKER, FILL, LINE
+    yield check_marker_properties, marker
+    yield check_fill_properties, marker
+    yield check_line_properties, marker
+    yield check_properties_existence, marker, MARKER, FILL, LINE, GLYPH
 
 def test_CircleX():
     marker = CircleX()
-    yield check_marker, marker
-    yield check_fill, marker
-    yield check_line, marker
-    yield check_props, marker, MARKER, FILL, LINE
+    yield check_marker_properties, marker
+    yield check_fill_properties, marker
+    yield check_line_properties, marker
+    yield check_properties_existence, marker, MARKER, FILL, LINE, GLYPH
 
 def test_Cross():
     marker = Cross()
-    yield check_marker, marker
-    yield check_fill, marker
-    yield check_line, marker
-    yield check_props, marker, MARKER, FILL, LINE
+    yield check_marker_properties, marker
+    yield check_fill_properties, marker
+    yield check_line_properties, marker
+    yield check_properties_existence, marker, MARKER, FILL, LINE, GLYPH
 
 def test_Diamond():
     marker = Diamond()
-    yield check_marker, marker
-    yield check_fill, marker
-    yield check_line, marker
-    yield check_props, marker, MARKER, FILL, LINE
+    yield check_marker_properties, marker
+    yield check_fill_properties, marker
+    yield check_line_properties, marker
+    yield check_properties_existence, marker, MARKER, FILL, LINE, GLYPH
 
 def test_DiamondCross():
     marker = DiamondCross()
-    yield check_marker, marker
-    yield check_fill, marker
-    yield check_line, marker
-    yield check_props, marker, MARKER, FILL, LINE
+    yield check_marker_properties, marker
+    yield check_fill_properties, marker
+    yield check_line_properties, marker
+    yield check_properties_existence, marker, MARKER, FILL, LINE, GLYPH
 
 def test_InvertedTriangle():
     marker = InvertedTriangle()
-    yield check_marker, marker
-    yield check_fill, marker
-    yield check_line, marker
-    yield check_props, marker, MARKER, FILL, LINE
+    yield check_marker_properties, marker
+    yield check_fill_properties, marker
+    yield check_line_properties, marker
+    yield check_properties_existence, marker, MARKER, FILL, LINE, GLYPH
 
 def test_Square():
     marker = Square()
-    yield check_marker, marker
-    yield check_fill, marker
-    yield check_line, marker
-    yield check_props, marker, MARKER, FILL, LINE
+    yield check_marker_properties, marker
+    yield check_fill_properties, marker
+    yield check_line_properties, marker
+    yield check_properties_existence, marker, MARKER, FILL, LINE, GLYPH
 
 def test_SquareCross():
     marker = SquareCross()
-    yield check_marker, marker
-    yield check_fill, marker
-    yield check_line, marker
-    yield check_props, marker, MARKER, FILL, LINE
+    yield check_marker_properties, marker
+    yield check_fill_properties, marker
+    yield check_line_properties, marker
+    yield check_properties_existence, marker, MARKER, FILL, LINE, GLYPH
 
 def test_SquareX():
     marker = SquareX()
-    yield check_marker, marker
-    yield check_fill, marker
-    yield check_line, marker
-    yield check_props, marker, MARKER, FILL, LINE
+    yield check_marker_properties, marker
+    yield check_fill_properties, marker
+    yield check_line_properties, marker
+    yield check_properties_existence, marker, MARKER, FILL, LINE, GLYPH
 
 def test_Triangle():
     marker = Triangle()
-    yield check_marker, marker
-    yield check_fill, marker
-    yield check_line, marker
-    yield check_props, marker, MARKER, FILL, LINE
+    yield check_marker_properties, marker
+    yield check_fill_properties, marker
+    yield check_line_properties, marker
+    yield check_properties_existence, marker, MARKER, FILL, LINE, GLYPH
 
 def test_X():
     marker = X()
-    yield check_marker, marker
-    yield check_fill, marker
-    yield check_line, marker
-    yield check_props, marker, MARKER, FILL, LINE
+    yield check_marker_properties, marker
+    yield check_fill_properties, marker
+    yield check_line_properties, marker
+    yield check_properties_existence, marker, MARKER, FILL, LINE, GLYPH
